Fixed: restore use reflection related code instead some random value (false in most case) Fixed: output result of loadpdump CLI command call.
git-svn-id: http://mangos.svn.sourceforge.net/svnroot/mangos@4314 26674162-90ff-0310-a85b-b64107c8dee9

+++ trunk/src/game/Object.cpp
@@ -198,10 +198,10 @@ void Object::BuildCreateUpdateBlockForPlayer(UpdateData *data, Player *target) c
     //sLog.outDebug("BuildCreateUpdate: update-type: %u, object-type: %u got flags: %X, flags2: %X", updatetype, m_objectTypeId, flags, flags2);
 
     ByteBuffer buf(500);
-    buf << updatetype;
+    buf << (uint8)updatetype;
     //buf.append(GetPackGUID());    //client crashes when using this
     buf << (uint8)0xFF << GetGUID();
-    buf << m_objectTypeId;
+    buf << (uint8)m_objectTypeId;
 
     _BuildMovementUpdate(&buf, flags, flags2);
